refactor(runtime): separate host input and driver lifecycle
